refactor: (step3) HangmanGameRoundInfo를 HangmanGameRoundTable의 정적 내부 클래스로 변경 HangmanGameRoundInfo는 HangmanGameRoundTable에서만 사용되는 객체이므로 탑레벨 클래스로 분리시키면 패키지구조가 복잡해질 것으로 우려
