+++ src/main/java/src/hangmanGame/HangmanGameRoundTable.java
@@ -24,4 +24,24 @@ public class HangmanGameRoundTable implements Iterable<Integer>{
     public Iterator<Integer> iterator() {
         return roundTable.keySet().iterator();
     }
+
+    static class HangmanGameRoundInfo {
+
+        private static final String INFO_PRINT_FORMAT = "남은 목숨 : %d, %s, 사용자 입력 : %c";
+
+        public final int remainingLife;
+        public final String correctingWord;
+        public final char inputAlphabet;
+
+        public HangmanGameRoundInfo(int remainingLife, String correctingWord, char inputAlphabet) {
+            this.remainingLife = remainingLife;
+            this.correctingWord = correctingWord;
+            this.inputAlphabet = inputAlphabet;
+        }
+
+        @Override
+        public String toString() {
+            return String.format(INFO_PRINT_FORMAT, remainingLife, correctingWord, inputAlphabet);
+        }
+    }
 }
